--- src/internal/interpreter/http/digest/index.ts
@@ -1,1 +1,0 @@
-export { DigestHelper } from './digest';
