+++ src/internal/interpreter/http/security/apiKey/index.ts
@@ -1,0 +1,1 @@
+export { ApiKeyHandler } from './api-key';
