+++ src/internal/interpreter/http/security/digest/digest.test.ts
@@ -1,12 +1,24 @@
+import {
+  DigestSecurityScheme,
+  DigestSecurityValues,
+  HttpScheme,
+  SecurityType,
+} from '@superfaceai/ast';
 import { createHash } from 'crypto';
+import { DEFAULT_AUTHORIZATION_HEADER_NAME } from '..';
+import { AuthCache } from '../../../../..';
 
-import { UnexpectedError } from '../../../errors';
-import { HttpResponse } from '../http';
-import { DigestHelper } from './digest';
+import { UnexpectedError } from '../../../../errors';
+import { HttpResponse } from '../../http';
+import { RequestContext } from '../interfaces';
+import { DigestHandler } from './digest';
 
-describe('DigestHelper', () => {
-  const mockUser = 'test-user';
-  const mockPassword = 'test-password';
+describe('DigestHandler', () => {
+  const id = 'digest';
+  const scheme = HttpScheme.DIGEST;
+  const type = SecurityType.HTTP;
+  const username = 'test-user';
+  const password = 'test-password';
   const mockUri = '/pms_api/91508/121/bookings/10619';
   const mockUrl = 'https://sky-eu1.clock-software.com' + mockUri;
   const mockRealm = 'API';
@@ -14,15 +26,59 @@ describe('DigestHelper', () => {
     'MTYzODM0OTE4Nzo0YTdlODlmYjI3ODZiZGZhNDhiODAwM2NmNjIwMzE2OQ==';
   const mockCnonce = '9e3355cb6a75d66ce7eea7fc7fc8526d';
   const mockOpaque = 'b816d4d26130bed8ccf5149e855000ad';
-  let mockInstance: DigestHelper;
+  let mockInstance: DigestHandler;
+  let configuration: DigestSecurityScheme & DigestSecurityValues;
+  let context: RequestContext;
+  let cache: AuthCache;
+  let retry: boolean;
 
   beforeEach(() => {
-    mockInstance = new DigestHelper(mockUser, mockPassword);
+    configuration = {
+      username,
+      password,
+      id,
+      scheme,
+      type,
+    };
+    context = {
+      pathParameters: {},
+      queryAuth: {},
+      headers: {},
+      requestBody: undefined,
+    };
+    cache = {};
   });
   afterEach(() => {
     jest.resetAllMocks();
   });
-  describe('extractCredentials', () => {
+  describe('prepare', () => {
+    it('does not change headers when cache is empty', () => {
+      mockInstance = new DigestHandler(configuration);
+      mockInstance.prepare(context, {});
+
+      expect(context.headers).toEqual({});
+    });
+
+    it('changes default authorization header when cache is not empty', () => {
+      mockInstance = new DigestHandler(configuration);
+      mockInstance.prepare(context, { digest: 'secret' });
+
+      expect(context.headers).toEqual({
+        [DEFAULT_AUTHORIZATION_HEADER_NAME]: 'secret',
+      });
+    });
+
+    it('changes custom authorization header when cache is not empty', () => {
+      configuration.authorizationHeader = 'custom';
+
+      mockInstance = new DigestHandler(configuration);
+      mockInstance.prepare(context, { digest: 'secret' });
+
+      expect(context.headers).toEqual({ custom: 'secret' });
+    });
+  });
+
+  describe('handle', () => {
     it('returns undefined on unexpected status code', () => {
       const qop = 'auth';
       const algorithm = 'MD5';
@@ -42,10 +98,16 @@ describe('DigestHelper', () => {
           },
         },
       };
+      retry = new DigestHandler(configuration).handle(
+        response,
+        mockUrl,
+        method,
+        context,
+        {}
+      );
 
-      expect(
-        mockInstance.extractCredentials(response, mockUrl, method)
-      ).toBeUndefined();
+      expect(retry).toEqual(false);
+      expect(context.headers).toEqual({});
     });
 
     it('throws on missing challenge header', async () => {
@@ -65,7 +127,13 @@ describe('DigestHelper', () => {
       };
 
       expect(() =>
-        mockInstance.extractCredentials(response, mockUrl, method)
+        new DigestHandler(configuration).handle(
+          response,
+          mockUrl,
+          method,
+          context,
+          {}
+        )
       ).toThrow(
         new UnexpectedError(
           `Digest auth failed, unable to extract digest values from response. Header "www-authenticate" not found in response headers.`
@@ -96,7 +164,13 @@ describe('DigestHelper', () => {
       };
 
       expect(() =>
-        mockInstance.extractCredentials(response, mockUrl, method)
+        new DigestHandler(configuration).handle(
+          response,
+          mockUrl,
+          method,
+          context,
+          {}
+        )
       ).toThrow(
         new UnexpectedError(
           `Digest auth failed, unable to extract digest values from response. Header "www-authenticate" does not contain scheme value eq. Digest`,
@@ -128,7 +202,13 @@ describe('DigestHelper', () => {
       };
 
       expect(() =>
-        mockInstance.extractCredentials(response, mockUrl, method)
+        new DigestHandler(configuration).handle(
+          response,
+          mockUrl,
+          method,
+          context,
+          {}
+        )
       ).toThrow(
         new UnexpectedError(
           `Digest auth failed, unable to extract digest values from response. Header "www-authenticate" does not contain "nonce"`,
@@ -159,7 +239,13 @@ describe('DigestHelper', () => {
       };
 
       expect(() =>
-        mockInstance.extractCredentials(response, mockUrl, method)
+        new DigestHandler(configuration).handle(
+          response,
+          mockUrl,
+          method,
+          context,
+          {}
+        )
       ).toThrow(
         new UnexpectedError(
           `Digest auth failed, parameter "algorithm" has unexpected value`,
@@ -190,7 +276,13 @@ describe('DigestHelper', () => {
       };
 
       expect(() =>
-        mockInstance.extractCredentials(response, mockUrl, method)
+        new DigestHandler(configuration).handle(
+          response,
+          mockUrl,
+          method,
+          context,
+          {}
+        )
       ).toThrow(
         new UnexpectedError(
           `Digest auth failed, parameter "quality of protection" has unexpected value`,
@@ -203,9 +295,11 @@ describe('DigestHelper', () => {
     it('prepares digest auth without qop and algorithm', async () => {
       const method = 'GET';
       const mockHeader = `Digest realm="${mockRealm}" nonce="${mockNonce}", opaque="${mockOpaque}"`;
+      mockInstance = new DigestHandler(configuration);
+
       //Prepare digest response
       const h1 = createHash('MD5')
-        .update(`${mockUser}:${mockRealm}:${mockPassword}`)
+        .update(`${username}:${mockRealm}:${password}`)
         .digest('hex');
       const h2 = createHash('MD5').update(`${method}:${mockUri}`).digest('hex');
       const digestResponse = createHash('MD5')
@@ -228,18 +322,68 @@ describe('DigestHelper', () => {
       };
 
       (mockInstance as any).makeNonce = () => mockCnonce;
+      retry = mockInstance.handle(response, mockUrl, method, context, cache);
 
-      expect(
-        mockInstance.extractCredentials(response, mockUrl, method)
-      ).toEqual(expect.stringContaining(`response="${digestResponse}"`));
+      expect(retry).toEqual(true);
+      expect(context.headers).toEqual({
+        [DEFAULT_AUTHORIZATION_HEADER_NAME]: expect.stringContaining(
+          `response="${digestResponse}"`
+        ),
+      });
+      expect(cache).toEqual({
+        digest: expect.stringContaining(`response="${digestResponse}"`),
+      });
+    });
+
+    it('prepares digest auth without qop and algorithm for custom authorization header', async () => {
+      const method = 'GET';
+      const mockHeader = `Digest realm="${mockRealm}" nonce="${mockNonce}", opaque="${mockOpaque}"`;
+      configuration.authorizationHeader = 'Custom';
+      mockInstance = new DigestHandler(configuration);
+
+      //Prepare digest response
+      const h1 = createHash('MD5')
+        .update(`${username}:${mockRealm}:${password}`)
+        .digest('hex');
+      const h2 = createHash('MD5').update(`${method}:${mockUri}`).digest('hex');
+      const digestResponse = createHash('MD5')
+        .update(`${h1}:${mockNonce}:${h2}`)
+        .digest('hex');
+
+      const response: HttpResponse = {
+        statusCode: 401,
+        body: 'HTTP Digest: Access denied.\n',
+        headers: {
+          'www-authenticate': mockHeader,
+        },
+        debug: {
+          request: {
+            url: mockUrl,
+            headers: {},
+            body: undefined,
+          },
+        },
+      };
+
+      (mockInstance as any).makeNonce = () => mockCnonce;
+      retry = mockInstance.handle(response, mockUrl, method, context, cache);
+
+      expect(retry).toEqual(true);
+      expect(context.headers).toEqual({
+        Custom: expect.stringContaining(`response="${digestResponse}"`),
+      });
+      expect(cache).toEqual({
+        digest: expect.stringContaining(`response="${digestResponse}"`),
+      });
     });
 
     it('prepares digest auth with auth-int qop', async () => {
       const qop = 'auth-int';
       const method = 'GET';
+      mockInstance = new DigestHandler(configuration);
 
       const h1 = createHash('MD5')
-        .update(`${mockUser}:${mockRealm}:${mockPassword}`)
+        .update(`${username}:${mockRealm}:${password}`)
         .digest('hex');
       const h2 = createHash('MD5').update(`${method}:${mockUri}`).digest('hex');
       const expectedResponse = createHash('MD5')
@@ -262,19 +406,25 @@ describe('DigestHelper', () => {
       };
 
       (mockInstance as any).makeNonce = () => mockCnonce;
+      retry = mockInstance.handle(response, mockUrl, method, context, cache);
 
-      expect(
-        mockInstance.extractCredentials(response, mockUrl, method)
-      ).toEqual(expect.stringContaining(`response="${expectedResponse}"`));
+      expect(retry).toEqual(true);
+      expect(context.headers).toEqual({
+        [DEFAULT_AUTHORIZATION_HEADER_NAME]: expect.stringContaining(`response="${expectedResponse}"`),
+      });
+      expect(cache).toEqual({
+        digest: expect.stringContaining(`response="${expectedResponse}"`),
+      });
     });
 
     it('prepares digest auth with default values and MD5', async () => {
       const qop = 'auth';
       const algorithm = 'MD5';
       const method = 'GET';
+      mockInstance = new DigestHandler(configuration);
 
       const h1 = createHash(algorithm)
-        .update(`${mockUser}:${mockRealm}:${mockPassword}`)
+        .update(`${username}:${mockRealm}:${password}`)
         .digest('hex');
       const h2 = createHash(algorithm)
         .update(`${method}:${mockUri}`)
@@ -299,22 +449,41 @@ describe('DigestHelper', () => {
       };
 
       (mockInstance as any).makeNonce = () => mockCnonce;
+      retry = mockInstance.handle(response, mockUrl, method, context, cache);
 
-      expect(
-        mockInstance.extractCredentials(response, mockUrl, method)
-      ).toEqual(expect.stringContaining(`response="${expectedResponse}"`));
+      expect(retry).toEqual(true);
+      expect(context.headers).toEqual({
+        [DEFAULT_AUTHORIZATION_HEADER_NAME]: expect.stringContaining(`response="${expectedResponse}"`),
+      });
+      expect(cache).toEqual({
+        digest: expect.stringContaining(`response="${expectedResponse}"`),
+      });
     });
 
-    it('prepares digest auth with default values and MD5-sess', async () => {
+    it('prepares digest auth with default values, custom headers and status code', async () => {
       const qop = 'auth';
-      const algorithm = 'MD5-sess';
+      const algorithm = 'MD5';
       const method = 'GET';
+      configuration.authorizationHeader = 'Auth';
+      configuration.challengeHeader = 'Challenge';
+      configuration.statusCode = 432;
+      mockInstance = new DigestHandler(configuration);
+
+      const h1 = createHash(algorithm)
+        .update(`${username}:${mockRealm}:${password}`)
+        .digest('hex');
+      const h2 = createHash(algorithm)
+        .update(`${method}:${mockUri}`)
+        .digest('hex');
+      const expectedResponse = createHash(algorithm)
+        .update(`${h1}:${mockNonce}:00000001:${mockCnonce}:${qop}:${h2}`)
+        .digest('hex');
 
       const response: HttpResponse = {
-        statusCode: 401,
+        statusCode: 432,
         body: 'HTTP Digest: Access denied.\n',
         headers: {
-          'www-authenticate': `Digest realm="${mockRealm}", qop="${qop}", algorithm=${algorithm}, nonce="${mockNonce}", opaque="${mockOpaque}"`,
+          'Challenge': `Digest realm="${mockRealm}", qop="${qop}", nonce="${mockNonce}", opaque="${mockOpaque}"`,
         },
         debug: {
           request: {
@@ -326,9 +495,25 @@ describe('DigestHelper', () => {
       };
 
       (mockInstance as any).makeNonce = () => mockCnonce;
+      retry = mockInstance.handle(response, mockUrl, method, context, cache);
+
+      expect(retry).toEqual(true);
+      expect(context.headers).toEqual({
+        'Auth': expect.stringContaining(`response="${expectedResponse}"`),
+      });
+      expect(cache).toEqual({
+        digest: expect.stringContaining(`response="${expectedResponse}"`),
+      });
+    });
+
+    it('prepares digest auth with default values and MD5-sess', async () => {
+      const qop = 'auth';
+      const algorithm = 'MD5-sess';
+      const method = 'GET';
+      mockInstance = new DigestHandler(configuration);
 
       let h1 = createHash('MD5')
-        .update(`${mockUser}:${mockRealm}:${mockPassword}`)
+        .update(`${username}:${mockRealm}:${password}`)
         .digest('hex');
       h1 = createHash('MD5')
         .update(`${h1}:${mockNonce}:${mockCnonce}`)
@@ -338,16 +523,6 @@ describe('DigestHelper', () => {
         .update(`${h1}:${mockNonce}:00000001:${mockCnonce}:${qop}:${h2}`)
         .digest('hex');
 
-      expect(
-        mockInstance.extractCredentials(response, mockUrl, method)
-      ).toEqual(expect.stringContaining(`response="${expectedResponse}"`));
-    });
-
-    it('prepares digest auth with default values and SHA-256', async () => {
-      const qop = 'auth';
-      const algorithm = 'SHA-256';
-      const method = 'GET';
-
       const response: HttpResponse = {
         statusCode: 401,
         body: 'HTTP Digest: Access denied.\n',
@@ -365,8 +540,25 @@ describe('DigestHelper', () => {
 
       (mockInstance as any).makeNonce = () => mockCnonce;
 
+      retry = mockInstance.handle(response, mockUrl, method, context, cache);
+
+      expect(retry).toEqual(true);
+      expect(context.headers).toEqual({
+        [DEFAULT_AUTHORIZATION_HEADER_NAME]: expect.stringContaining(`response="${expectedResponse}"`),
+      });
+      expect(cache).toEqual({
+        digest: expect.stringContaining(`response="${expectedResponse}"`),
+      });
+    });
+
+    it('prepares digest auth with default values and SHA-256', async () => {
+      const qop = 'auth';
+      const algorithm = 'SHA-256';
+      const method = 'GET';
+      mockInstance = new DigestHandler(configuration);
+
       const h1 = createHash('sha256')
-        .update(`${mockUser}:${mockRealm}:${mockPassword}`)
+        .update(`${username}:${mockRealm}:${password}`)
         .digest('hex');
       const h2 = createHash('sha256')
         .update(`${method}:${mockUri}`)
@@ -375,16 +567,6 @@ describe('DigestHelper', () => {
         .update(`${h1}:${mockNonce}:00000001:${mockCnonce}:${qop}:${h2}`)
         .digest('hex');
 
-      expect(
-        mockInstance.extractCredentials(response, mockUrl, method)
-      ).toEqual(expect.stringContaining(`response="${expectedResponse}"`));
-    });
-
-    it('prepares digest auth with default values and SHA-256-sess', async () => {
-      const qop = 'auth';
-      const algorithm = 'SHA-256-sess';
-      const method = 'GET';
-
       const response: HttpResponse = {
         statusCode: 401,
         body: 'HTTP Digest: Access denied.\n',
@@ -402,8 +584,25 @@ describe('DigestHelper', () => {
 
       (mockInstance as any).makeNonce = () => mockCnonce;
 
+      retry = mockInstance.handle(response, mockUrl, method, context, cache);
+
+      expect(retry).toEqual(true);
+      expect(context.headers).toEqual({
+        [DEFAULT_AUTHORIZATION_HEADER_NAME]: expect.stringContaining(`response="${expectedResponse}"`),
+      });
+      expect(cache).toEqual({
+        digest: expect.stringContaining(`response="${expectedResponse}"`),
+      });
+    });
+
+    it('prepares digest auth with default values and SHA-256-sess', async () => {
+      const qop = 'auth';
+      const algorithm = 'SHA-256-sess';
+      const method = 'GET';
+      mockInstance = new DigestHandler(configuration);
+
       let h1 = createHash('sha256')
-        .update(`${mockUser}:${mockRealm}:${mockPassword}`)
+        .update(`${username}:${mockRealm}:${password}`)
         .digest('hex');
       h1 = createHash('sha256')
         .update(`${h1}:${mockNonce}:${mockCnonce}`)
@@ -415,9 +614,32 @@ describe('DigestHelper', () => {
         .update(`${h1}:${mockNonce}:00000001:${mockCnonce}:${qop}:${h2}`)
         .digest('hex');
 
-      expect(
-        mockInstance.extractCredentials(response, mockUrl, method)
-      ).toEqual(expect.stringContaining(`response="${expectedResponse}"`));
+      const response: HttpResponse = {
+        statusCode: 401,
+        body: 'HTTP Digest: Access denied.\n',
+        headers: {
+          'www-authenticate': `Digest realm="${mockRealm}", qop="${qop}", algorithm=${algorithm}, nonce="${mockNonce}", opaque="${mockOpaque}"`,
+        },
+        debug: {
+          request: {
+            url: mockUrl,
+            headers: {},
+            body: undefined,
+          },
+        },
+      };
+
+      (mockInstance as any).makeNonce = () => mockCnonce;
+
+      retry = mockInstance.handle(response, mockUrl, method, context, cache);
+
+      expect(retry).toEqual(true);
+      expect(context.headers).toEqual({
+        [DEFAULT_AUTHORIZATION_HEADER_NAME]: expect.stringContaining(`response="${expectedResponse}"`),
+      });
+      expect(cache).toEqual({
+        digest: expect.stringContaining(`response="${expectedResponse}"`),
+      });
     });
   });
 });
