+++ src/internal/interpreter/http/security/http/index.ts
@@ -1,0 +1,1 @@
+export { HttpHandler } from './http';
